[BOL-375-4] repo impl refactor (#130)
* BOL-375 refactor: GroupSearchUiModel data package로 이동

* BOL-375 refactor: ic_owner 제거, designsystem manager로 변경

* BOL-375 refactor: 중복 icon 제거

* BOL-375 refactor: mock img 제거

* BOL-375 refactor: 필요 없는 png 제거

* BOL-375 refactor: guest user의 dice vector 프레젠테이션 모듈에서 디자인시스템으로 이동

* BOL-375 refactor: ic_x 제거 (디자인시스템에 존재)

* BOL-375 refactor: ic_plus 제거 (디자인시스템에 존재)

* BOL-375 refactor: ic_gallery 제거 (디자인시스템에 ic_image 존재)

* BOL-375 refactor: ic_empty_dice 제거 (디자인시스템에 img_dice_empty 존재)

* BOL-375 refactor: ic_time 제거 (디자인시스템에 ic_time_line_16 존재)

* BOL-375 refactor: 사용하지 않는 mipmap 파일 제거

* BOL-375 feat: force upgrade play store 주소 변경

* BOL-375 feat: group search activity test 제거

* BOL-375 refactor: auth VM package 이동 (viewmodel.login -> view.login.auth)

* BOL-375 refactor: ic_calendar_line_16으로 바꿔주고 리팩토링

* BOL-375 refactor: image view에 app tint 사용 가능하도록 lint ignore (api 21 이상부터는 발생하지 않는 문제라 ignore)

* BOL-375 refactor: ic_time_line_16으로 변경

* BOL-375 refactor: move AuthRepositoryImpl's package

* BOL-375 refactor: 사용하지 않는 모델 제거

* BOL-375 refactor: rename ImageFileApi -> FileApi

* BOL-375 refactor: api 서버 맞춰서 추가

* BOL-375 refactor: Mapper renaming

* BOL-375 refactor: move Auth Datasource's package

* BOL-375 refactor: DataSource Impl package 제거

* BOL-375 refactor: LoginType package 이동

* BOL-375 refactor: RankDataSource import문 제거

* BOL-375 feat: MemberDataSource memberApi 맞춰서 구현

* BOL-375 feat: FileDataSource fileApi 맞춰서 구현

* BOL-375 feat: GameDataSource GameApi 맞춰서 구현

* BOL-375 feat: MatchDataSource MatchApi 맞춰서 구현

* BOL-375 feat: TermsDataSource TermsApi 맞춰서 구현

* BOL-375 feat: UserDataSource UserApi 맞춰서 구현

* BOL-375 refactor: fileDataSource로 FileKey constant 이동

* BOL-375 feat: GroupDataSource GroupApi 맞춰서 구현

* BOL-375 feat: AuthDataSource AuthApi 맞춰서 구현

* BOL-375 feat: unuse dataSource files remove

* BOL-375 refactor: create group api fun 이름 맞지 않는 것 수정

* BOL-375 refactor: domain repository 서버 맞춰서 리팩토링

* BOL-375 refactor: unuse repository remove

* BOL-375 refactor: UseCase refactoring한 Repository로 변경

* BOL-375 refactor: Mapper 서버 맞춰서 리팩터링

* BOL-375 refactor: Repository Impl 파일들 서버 맞춰서 리팩토링

* BOL-375 refactor: BaseItem 제거를 위한 ErrorItem으로 refactoring

* BOL-375 refactor: BaseResponse -> ErrorResponse로 renaming

* BOL-375 refactor: BaseItem remove
