+++ data/src/main/java/com/yapp/bol/data/mapper/CoreMapper.kt
@@ -1,0 +1,15 @@
+package com.yapp.bol.data.mapper
+
+import com.yapp.bol.data.model.base.ErrorResponse
+import com.yapp.bol.domain.model.ApiResult
+import com.yapp.bol.domain.model.ErrorItem
+
+object CoreMapper {
+
+    fun ApiResult<ErrorResponse>.mapperToBaseItem(): ApiResult<ErrorItem> {
+        return when (this) {
+            is ApiResult.Success -> ApiResult.Success(ErrorItem(data.code, data.message))
+            is ApiResult.Error -> ApiResult.Error(exception)
+        }
+    }
+}
