+++ data/src/main/java/com/yapp/bol/data/mapper/FileMapper.kt
@@ -1,0 +1,14 @@
+package com.yapp.bol.data.mapper
+
+import com.yapp.bol.data.model.group.response.ImageFileUploadResponse
+import com.yapp.bol.domain.model.ApiResult
+
+object FileMapper {
+
+    fun ApiResult<ImageFileUploadResponse>.fileUploadToDomain(): ApiResult<String> {
+        return when (this) {
+            is ApiResult.Success -> ApiResult.Success(data.url)
+            is ApiResult.Error -> ApiResult.Error(exception)
+        }
+    }
+}
