+++ data/src/main/java/com/yapp/bol/data/model/base/ErrorResponse.kt
@@ -1,6 +1,6 @@
 package com.yapp.bol.data.model.base
 
-data class BaseResponse(
+data class ErrorResponse(
     val code: String = "",
     val message: String = "",
 )
