--- domain/src/main/java/com/yapp/bol/domain/model/BaseItem.kt
@@ -1,6 +1,0 @@
-package com.yapp.bol.domain.model
-
-data class BaseItem(
-    val code: String,
-    val message: String,
-)
